Fix: multiplayer starting coordinates were wrong
git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@6708 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ TileEngine/Map Edgepoints.cpp
@@ -118,8 +118,8 @@ void FillCentreGridnos(BOOLEAN fCenterOnly)
 	INT32 y_base	= world_rows_adjusted/2;
 
 	INT32 totallen  = std::sqrt((float)(world_rows_adjusted * world_rows_adjusted + world_cols_adjusted * world_cols_adjusted));
-	INT32 droplen   = totallen * (1.0f - 2 * xratio);
-	INT32 steplen   = droplen / CENTER_ENTRY_POINTS_ROW;
+	INT32 droplen   = (FLOAT)(totallen * (1.0f - 2 * xratio));
+	FLOAT steplen   = (FLOAT)(droplen) / (FLOAT)(CENTER_ENTRY_POINTS_ROW);
 
 	UINT16 cnt = 0;
 	for(UINT16 i = 0; i < CENTER_ENTRY_POINTS_ROW; ++i)
